+++ encoding/thrift/observability_test.go
@@ -180,7 +180,7 @@ func TestThriftMetrics(t *testing.T) {
 					{Name: "timeout_ttl_ms"},
 					{Name: "ttl_ms", Value: []int64{1000}},
 				}
-				testutils.AssertHistograms(t, wantHistograms, serverMetricsRoot.Snapshot().Histograms)
+				testutils.AssertClientAndServerHistograms(t, wantHistograms, clientMetricsRoot, serverMetricsRoot)
 			})
 		})
 	}