+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.49.0] - 2020-10-23
 ### Added
 - observability: Add request and response payload size histogram
 - api: expose `BodySize` field in the transport request/response, middleware can use
@@ -1299,7 +1299,8 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
-[Unreleased]: https://github.com/yarpc/yarpc-go/compare/v1.48.0...HEAD
+[Unreleased]: https://github.com/yarpc/yarpc-go/compare/v1.49.0...HEAD
+[1.49.0]: https://github.com/yarpc/yarpc-go/compare/v1.48.0...v1.49.0
 [1.48.0]: https://github.com/yarpc/yarpc-go/compare/v1.47.2...v1.48.0
 [1.47.2]: https://github.com/yarpc/yarpc-go/compare/v1.47.1...v1.47.2
 [1.47.1]: https://github.com/yarpc/yarpc-go/compare/v1.47.0...v1.47.1